all suggestions done, refactoring folders and splitting massive files
